fix(core): change breakpoints for demo experience
